RF: Move export_archive() out of plugins
Towards a fix for #4519
